Remove unnecessary mod class abstraction
